Revert "Restore incomplete string and macro functionality"
This reverts commit dd855cbc25f5acab77175c86f8a43225651da712.
