+++ Cargo.toml
@@ -21,22 +21,14 @@ name = "twofloat"
 test = true
 
 [features]
-default = ["math_funcs", "string_convert"]
+default = ["math_funcs"]
 math_funcs = ["num-traits/std"]
-string_convert = ["lazy_static", "num-bigint", "num-rational"]
 
 [dependencies]
 hexf = "0.2"
-lazy_static = { version = "1.4", optional = true }
-num-bigint = { version = "0.3", optional = true }
-num-rational = { version = "0.3", optional = true }
 num-traits = { version = "0.2.14", default-features = false }
 serde = { version = "1.0", default-features = false, features = ["derive"], optional = true }
 
 [dev-dependencies]
 rand = "0.8"
 serde_test = "1.0"
-
-[workspace]
-members = ["twofloat_macro"]
-default-members = [".", "twofloat_macro"]
