+++ src/functions/explog.rs
@@ -1,7 +1,6 @@
 use hexf::hexf64;
 
-use crate::consts::LN_2;
-use crate::TwoFloat;
+use crate::{consts::LN_2, TwoFloat};
 
 // 1/ln(2)
 const FRAC_1_LN_2: TwoFloat = TwoFloat {
