+++ src/functions/fraction.rs
@@ -1,5 +1,4 @@
-use crate::arithmetic::fast_two_sum;
-use crate::TwoFloat;
+use crate::{arithmetic::fast_two_sum, TwoFloat};
 
 impl TwoFloat {
     /// Returns the fractional part of the number.
