+++ tests/arithmetic_tests.rs
@@ -1,7 +1,6 @@
 #![allow(clippy::float_cmp)]
 
-use core::convert::TryFrom;
-use core::fmt::Debug;
+use core::{convert::TryFrom, fmt::Debug};
 
 use rand::Rng;
