--- twofloat_macro/Cargo.toml
@@ -1,24 +1,0 @@
-[package]
-name = "twofloat_macro"
-version = "0.5.0"
-authors = ["Andrew Tribick"]
-
-keywords = ["float", "precision", "numerics", "floating-point", "arithmetic"]
-categories = ["algorithms", "mathematics", "science"]
-edition = "2018"
-
-readme = "README.md"
-license = "BSD-3-Clause"
-description = "Macros for double-double arithmetic functionality."
-repository = "https://github.com/ajtribick/twofloat"
-
-[lib]
-name = "twofloat_macro"
-proc-macro = true
-
-[dependencies]
-syn = { version = "1.0.82", default-features = false, features = ["parsing", "proc-macro"] }
-twofloat = { version = "0.5.0", path = ".." }
-
-[dev-dependencies]
-hexf = "0.2"
